Remove Abstract class and rename the shared methods to the original names
