+++ kroxylicious-kms-provider-kroxylicious-inmemory-test-support/src/main/java/io/kroxylicious/kms/provider/kroxylicious/inmemory/InMemoryTestKekManager.java
@@ -6,11 +6,14 @@
 
 package io.kroxylicious.kms.provider.kroxylicious.inmemory;
 
+import java.util.Objects;
 import java.util.UUID;
+import java.util.concurrent.CompletionException;
 
-import io.kroxylicious.kms.service.AbstractTestKekManager;
+import io.kroxylicious.kms.service.TestKekManager;
+import io.kroxylicious.kms.service.UnknownAliasException;
 
-public class InMemoryTestKekManager extends AbstractTestKekManager {
+public class InMemoryTestKekManager implements TestKekManager {
     private final InMemoryKms kms;
 
     public InMemoryTestKekManager(InMemoryKms kms) {
@@ -18,25 +21,53 @@ public class InMemoryTestKekManager extends AbstractTestKekManager {
     }
 
     @Override
-    protected UUID read(String alias) {
+    public UUID read(String alias) {
         return kms.resolveAlias(alias).toCompletableFuture().join();
     }
 
     @Override
-    protected void create(String alias) {
-        var kekId = kms.generateKey();
-        kms.createAlias(kekId, alias);
+    public void generateKek(String alias) {
+       Objects.requireNonNull(alias);
+
+        try {
+            read(alias);
+            throw new AlreadyExistsException(alias);
+        }
+        catch (CompletionException e) {
+            if (e.getCause() instanceof UnknownAliasException) {
+                var kekId = kms.generateKey();
+                kms.createAlias(kekId, alias);
+            }
+            else {
+                throw unwrapRuntimeException(e);
+            }
+        }
     }
 
     @Override
-    protected void delete(String alias) {
-        var kekRef = read(alias);
-        kms.deleteAlias(alias);
-        kms.deleteKey(kekRef);
+    public void deleteKek(String alias) {
+        Objects.requireNonNull(alias);
+        try {
+            var kekRef = read(alias);
+            kms.deleteAlias(alias);
+            kms.deleteKey(kekRef);
+        }
+        catch (CompletionException e) {
+            throw unwrapRuntimeException(e);
+        }
     }
 
     @Override
-    protected void rotate(String alias) {
-        create(alias);
+    public void rotateKek(String alias) {
+        Objects.requireNonNull(alias);
+
+        try {
+            read(alias);
+            var kekId = kms.generateKey();
+            kms.createAlias(kekId, alias);
+        }
+        catch (CompletionException e) {
+            throw unwrapRuntimeException(e);
+        }
     }
 }
